Show context info along with progress
This commit is a quick and dirty mockup where, along with the
progress, we also inform users of what the app is doing.

I implemented it in a ghetto way by increasing the height of a cell
and making the progress bar always visible.

It is not visually pleasing but it shows the point that I
wanted to make, i.e. that adding a little context is nicer
because one better understands what is happening.

Thanks to @lorenzoPrimi who helped me to understand what was
the best and quickest way to implement this change :-).